Use generic vc functions for both SVN and Git
eide-vc.el was designed to support a project registered in both SVN and
Git. In practice, it never happens. SVN and Git are usually mixed with
"git svn" commands, which results in a Git only repository.

The code can be simplified if we don't need to check all VC backends
anymore.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>