Improve language support in the menu
The symbols in the menu are parsed by imenu. But the imenu function
that was called (imenu--generic-function) was not appropriate. It seems
to be a fallback method.

It was used because initially cc-imenu-c-generic-expression and
cc-imenu-c++-generic-expression were overriden. This was removed in
version 1.9.

Now the proper way to do it is to use imenu--index-alist.

Many more languages are supported (Perl, Python, ...).

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>