Don't open a project unless explicitly requested
When Emacs is launched, if a project is defined in the current
directory, don't open it, unless the new option --eide-op (op = open
project) is given.

When changing the root directory, if a project is defined in the new
directory, don't open it, and don't even clear the desktop.

A new command appears in the menu when a project is defined in the
current root directory: 'Load the project present in this directory'.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>