Create a project without symbols
Add the possibility to create a project without tags/cscope symbols.

It can be useful to disable symbols, if the source tree is a build
system containing several source trees (packages), for which you don't
need symbols, and which would result in huge and useless databases.

If symbols are disabled for the current project:
- Disable tags/cscope key bindings and menu commands
- Stop the creation of tags and cscope list of files if not yet
  finished
- Remove tags and cscope list of files

Show global grep command in popup menu when some text is selected, even
if no project is defined.

Update the help page.

Remove unused function eide-project-get-config-value.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>