Fix unwanted selection of a project in the list
In the list of projects, prevent the user from selecting a project by
clicking in the empty space on the right.
The clickable properties of the project name currently extends over the
whole line. An extra space must be added after each project name to
avoid this. The same trick is already used in the menu buffer.

The lists of projects are updated as soon as they are accessed.

The extra space could have been added at display time only, to avoid
modifying the format of the projects list file. But the user may remove
a project from the list with the popup menu, and in that case the
buffer must be saved.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>