Replace path in compilation output on the fly
When using "compile_error_old_path_regexp" option in project
configuration, instead of replacing the path in the whole compilation
buffer when it is finished, it can be replaced on the fly, with
compilation-filter-hook.

With the old method, in case of parallel make, the compilation buffer
could be displaying errors while the compilation was not finished, and
the user could click on a link that was not updated yet.

Remove the check of the existence of compilation-filter-hook, because
we are already checking that Emacs version >= 24.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>