Use a single window for completions
When switching from display-buffer-function to display-buffer-alist,
the way "*Completions*" buffer is displayed was modified. Instead of
using "output" window when visible (and enlarging it if necessary), the
code was simplified to systematically create a new window (half the
size of the frame) at bottom.

The drawback of this method is that several windows can be created for
completions, when called several times within a single find-file
command.

Before creating a new window, we must check if it is possible to reuse
an existing window already displaying "*Completions*" buffer.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>