Fix the consistency of IDE windows display
The windows layout can often be inconsistent, either because IDE
windows are displayed twice, or because the internal window objects are
not valid anymore.
It can happen when Emacs restores an old window configuration. In
particular, find-file command saves the window configuration, and
restores it when the command is finished - either normally or cancelled
by C-g - whatever happened in between (including hiding or showing IDE
windows).

To avoid that, we can register a hook on every window configuration
change. The purpose of this hook is to check the internal information
(the hide/show status and the validity of window objects) and make sure
that it is consistent with the display. Both internal information and
the display can be updated to fix such an issue.

Now the name of the buffer displayed in "output" window is always
updated - not only when hiding IDE windows - because the hook will
check its visibility.

The size of the output buffer should also be saved before it is resized
by completions.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>