Replace "touch" with the equivalent in Emacs Lisp
To create a file, or update its modification time, "touch" shell
command is called.

It is more appropriate to replace it with the equivalent in Emacs Lisp.

In eide-project-create, the creation of the file is removed because it
is useless.

The eide-option-touch-files-when-using-flag option is removed, and
files will always be updated, so that they are recompiled for sure (in
case of source code). I can't think of a good reason to disable this
behaviour.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>