Keep buffer up-to-date when switching to .ref/.new file
When switching to .ref/.new file, the buffer is considered to be
out-of-date, and Emacs suggests that it should be updated, when the
user tries to modify it.

This issue was introduced when replacing "touch" shell command with its
equivalent in Emacs Lisp.

The revert-buffer function must be called after set-file-times
function.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>