Use standard Emacs Lisp indentation (if)
In Emacs Lisp, in "if" test, the "then" block must be more indented
than the "else" block.

This specific indentation rule for "if" is currently not respected,
because I'm used to other languages that have the same indentation
level for "then" and "else" blocks, and I'm not really at ease with
this particularity.

However, the change is applied to the whole source code, to be
compliant with standard Emacs Lisp coding rules.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>