Emacs fails to exit when a process is running
When trying to exit while a process (shell, grep...) is running, Emacs
reuse the "output" window to display "*Process List*" buffer.
In eide-windows-configuration-change-hook, we detect that the expected
"output" buffer is not displayed. We believe that the "output" window
was closed, and we close all IDE windows.
Emacs gets lost because it is trying to access the "output" window that
we have just closed.

The content of the "output" window should not be strictly checked,
because we're not always notified. Instead, we should just check that:
* the "menu" window is displayed in a consistent window object
* the "output" window is consistent (whatever the content)

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>