Fix broken layout when exiting dired mode
When either exiting from dired mode with mouse-3, without selecting a
file, or changing root directory (which calls dired), the windows
layout gets broken.

In eide-menu-browsing-mode-stop, one of the first actions is to call
eide-windows-skip-unwanted-buffers-in-source-window. In this function,
for every bury-buffer call, eide-windows-configuration-change-hook is
also called. When it comes to the menu buffer, it results in showing
IDE windows (eide-windows-show-ide-windows).
The problem is that the layout had been saved in
eide-menu-browsing-mode-start (eide-windows-save-and-unbuild-layout),
and will be restored in eide-menu-browsing-mode-stop
(eide-windows-restore-layout).

Disable eide-windows-configuration-change-hook in
eide-windows-skip-unwanted-buffers-in-source-window.

In eide-menu-browsing-mode-stop, restore the layout before calling
eide-windows-skip-unwanted-buffers-in-source-window.

In eide-project-change-root, there is no need to temporary hide the IDE
windows, because desktop-change-dir is not called.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>