Fix buffer list when buffers are renamed
When two files with the same name are opened, buffers are suffixed with
their directory name. But if one of the buffer is closed, the suffix is
removed from the other buffer name. As a consequence,
eide-menu-files-list contains a non existent buffer name, and the
interface is broken.

The buffer list is already updated when a new file is opened. So the
problem is only when a buffer is closed.

Update eide-menu-files-list whenever a buffer with <...> suffix is
closed.

Signed-off-by: Cédric Marie <cedric.marie@openmailbox.org>