Prevent source window from displaying the menu
When selecting a project with no open file, the source window displays
the menu buffer.

Always make sure that the buffer displayed in the source window is
correct, when restoring IDE windows.

In eide-windows-show-ide-windows:

* Call eide-windows-skip-unwanted-buffers-in-source-window
* Don't call eide-menu-update anymore, since it will be called by
  eide-windows-skip-unwanted-buffers-in-source-window.

In eide-windows-show-ide-windows and eide-windows-hide-ide-windows,
restore window-configuration-change-hook before calling
eide-windows-skip-unwanted-buffers-in-source-window, because this
function also needs to temporary remove this hook.

Signed-off-by: Cédric Marie <cedric@hjuvi.lautre.net>