Prevent Emacs from loading a global desktop
When loading a project at startup (with --eide-op option), if a desktop
file is present in ~ or ~/.emacs.d, it will be loaded instead of the
one of the project.

To prevent this, it is necessary to set desktop-path to the root
directory of the project. (Its default value is a list containing ~ and
~/.emacs.d). It should not be set to nil either, because in that case,
a desktop file will also be searched in ~/.emacs.d.

It is only necessary at startup. If the project is loaded later on,
desktop-read and desktop-change-dir functions are called with
eide-root-directory as an argument.

There is no need to modify desktop-dirname variable. It is not used to
find and load the desktop. It is an internal variable that is
automatically updated when the desktop is read, in order to save the
desktop file later on.

In eide-project-change-root, use desktop-dirname variable to save the
previous desktop, instead of creating another variable
(l-old-root-directory).

Signed-off-by: Cédric Marie <cedric@hjuvi.lautre.net>