Add a changelog file
Create CHANGELOG.rst, to describe the changes between versions.

The information was taken from the "Download" page of the website. It
has been a little bit reworked to mostly comply with
keepachangelog.com (using Added, Changed, Removed, and Fixed types).
A Performance type is also used, because these kinds of change do not
fit in the other categories (Such a type is used by GitLab CE for
example).

NB: The "Download" page will be removed from the website.

Signed-off-by: Cédric Marie <cedric@hjuvi.lautre.net>