Convert reStructuredText files to Markdown
Convert reStructuredText files (README.rst, AUTHORS.rst, and
CHANGELOG.rst) to Markdown (.md).

Markdown is supported by all hosting services, while reStructuredText
is not always supported, or at least not natively (Gitea requires an
external renderer).
Although reStructuredText is more appropriate for complex
documentation, Markdown can be enough for simple files.

Integrate some information from the website, so that the detailed
instructions are available in the source code, and the website can be
kept as simple as possible, and refer to the README file.
The purpose is also to avoid the duplication of the information,
between the source code and the website.

Remove INSTALL.rst file (the instructions are provided in the README.md
file directly).

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>