Remove the "Help" buffer
Remove the "Help" buffer which provided some instructions directly inside
Emacs. Instead, merge the instructions into the README file, which contained a
large part of the information already, by adding the missing details.

This avoids embedding too much text in the package at run-time, and having
duplicate information.

Replace it with a simple "About" popup message (with the version and the
URL of the homepage).

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>