Browse through cscope results with F7/F8
F7/F8 keys (previous/next) only work with grep results and
errors/warnings in the compilation output.

Make them work with cscope results as well.

Depending on the version of xcscope.el, the name of the functions to
call can be different (cscope-history-backward-line and
cscope-history-forward-line vs. cscope-prev-symbol and
cscope-next-symbol), so we must check which one is defined with fboundp
(and we must check them both to avoid warnings when installing the
package with the two possible versions).

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>