Make sure cscope buffer name is always restored
When reaching the beginning or the end of the cscope result list, with
F7/F8, the cscope function exits with an error, and the rest of our
function, which restores the cscope buffer name, is not executed.

We must protect such cscope calls with unwind-protect, so that we can clean
up and restore the cscope buffer name, whether the cscope call was
successful or not.

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>