Keep open files when changing the workspace
When switching to another workspace, the current project (if any) is
closed, and all open files are closed (the desktop is cleared).

This makes sense because a project should not be part of two
workspaces, so when we're switching to a new workspace, we're not
supposed to keep on working on this project.

But if we just opened some files, and decide to create a new project
with these files, in a specific workspace, then we don't want the files
to be closed when we switch to the desired workspace.

So keep the files open, if there is no project currently loaded, when
switching to another workspace.

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>