Load projects in their dedicated workspace
When loading a project with --eide-op option, the project is always
added in the first workspace, while it may already be part of another
one. When loading it with "Load the project present in this directory"
in the menu, the project is added to the current workspace.

In all cases, if the project is already part of a workspace, we must
switch to this workspace before loading the project.

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>