Add Cscope support for Java files
The Cscope project homepage says: "The fuzzy parser supports C, but is
flexible enough to be useful for C++ and Java".

Add support for .java files, in addition to C/C++ files.

NB: To avoid worrying about source file type, I could use "cscope -bR"
command, but it seems to be parsing only C/C++ files (not Java), and it
would break the filtering feature (cscope_exclude_files and
cscope_exclude_dirs in the project configuration).

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>