Add an option for Cscope to support extra languages
Cscope was designed for C language, but is suitable for many other
languages. The homepage mentions C++ and Java, but it also seems to
support Python, Rust and Go for example.

The recursive command to build the database ("cscope -bR") only parses
C/C++ files, so it is really necessary to build the list of files
(cscope.files). Building this list is also required for the filtering
feature (cscope_exclude_files and cscope_exclude_dirs in the project
configuration).

The previous commit added support for Java language. But instead of
hard-coding specific extra languages, let's add an option (in the
customization: "Code browsing > List of Cscope extra supported file
extensions") so that the user can add whatever language he/she wants.

The C and C++ languages are still hard-coded, because they are the
basic usage of Cscope, and the default value of the new option includes
.java (Java), .py (Python), .rs (Rust) and .go (Go).

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>