Don't create empty result buffer with old name
When switching to another project, all the result buffers are closed by
desktop-clear. But the value of eide-windows-output-window-buffer is
still the name of the previously displayed result buffer. And
eide-windows-show-ide-windows will create an empty result buffer with
this name by calling switch-to-buffer.

So, when choosing the output buffer to display in
eide-windows-show-ide-windows, make sure the buffer exists. Otherwise
display the default "*results*" buffer (which is created if it doesn't
exist).

To avoid closing and creating this default buffer every time, add
"*results*" in desktop-clear-preserve-buffers list.

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>