Require Emacs 26.1 directly in eide-pkg.el
Change the minimum version of Emacs required from 25.1 to 26.1
(released in 2018), since this is the oldest version I can really check
currently.

Add the dependency on Emacs 26.1 directly in eide-pkg.el, and not only
in user-install.sh script, so that it can also be checked with standard
package installation.
Remove the check from user-install.sh script.

Also add the homepage in the package definition.

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>