Make sure to remove old versions of the package
In user-install.sh, before creating the package, make sure to remove
all possible old versions of the local package file, and all possible
old installed versions (in ~/.emacs.d/elpa). This is useful in case of
version upgrade.

In user-uninstall.sh, make sure to also remove old package files
(although this should not be necessary with the previous change), and
all possible installed versions in ~/.emacs.d/elpa.

In both scripts, check that it is launched from the root directory (all
paths are relative).

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>