Clean the workspace as a separate action
When displaying the list of projects in the current workspace, we check
the existence of each directory and suggest to remove the project if
the directory doesn't exist anymore. This was introduced in version
2.3.2.

There are several drawbacks:
* It takes longer to display if disk I/O are busy.
* If some projects are located in an external storage that is currently
  removed, it will ask a confirmation for each of them (and clicking on
  a moving yes/no popup is really annoying).

Instead, let's create a separate action to clean the list. It also
displays the list, but when the cleaning is finished, it doesn't let
you select a project, but switches back to the editor mode instead.

Add a separator for workspace edition actions in the popup menu (Add
project / Remove project / Clean workspace).

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>