Rewrite the description of the package
The current description is unclear about the IDE features provided by
this package, since Emacs is already considered to be an IDE itself.

The IDE features are not "provided" by the package, but "made available
out of the box". It is important to explain that the usage of these
features is automated and integrated directly into the interface.

Add an "Information" section header for the technical details, because
it is a list, and it gets mixed with the description that also ends
with a list.

Signed-off-by: Cédric Marie <cedric@hjuvi.fr.eu.org>