Remove "specific background color" option
Remove eide-custom-menu-use-specific-background-color option in
eide-menu group.

This option was useful in the beginning with eide-light color theme, to
have a white background for the menu, while the source code and the
results had an "old lace" background.

But with more recent versions of Emacs, the background color properties
stopped applying on the whole lines of the menu buffer. It applied only
on the text itself, but not beyond the new line character. It was no
longer the expected effect, and made the menu look weird.

So remove the option and simplify the code. The background color of the
menu is now "old lace" (like other buffers) with eide-light theme.

Remove eide-i-menu-insert-text function because it only calls insert
when the option is removed.